requirements.txt
autopep8==1.5
pre-commit==2.2.0
pycodestyle==2.5.0
pycryptodome==3.9.7
pyflakes==2.1.1
tqdm==4.46.0
